requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
python-multipart
Pillow
opencv-python-headless
numpy
